Removed directories from example project and improved README
