+++ project/src/com/google/daggerquery/example/Bed.java
@@ -14,20 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-package com.google.daggerquery.example.models;
+package com.google.daggerquery.example;
 
 import javax.inject.Inject;
-import java.util.Set;
 
-public class Beach {
-  private final Set<Tourist> tourists;
-  private final Set<Staff> lifeguards;
-  private final IceCreamShop iceCreamShop;
-
+public class Bed {
   @Inject
-  Beach(Set<Tourist> tourists, Set<Staff> lifeguards, IceCreamShop iceCreamShop) {
-    this.iceCreamShop = iceCreamShop;
-    this.tourists = tourists;
-    this.lifeguards = lifeguards;
+  Bed() {
   }
 }
