+++ project/src/com/google/daggerquery/example/LifeVest.java
@@ -14,25 +14,32 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-package com.google.daggerquery.example.models;
+package com.google.daggerquery.example;
 
 import javax.inject.Inject;
+import java.util.Random;
 
-public class Villa {
-  private final Bed bed;
-  private final MiniBar miniBar;
-  private final TV tv;
-  private final Phone phone;
-  private final Pool pool;
-  private final Tourist tourist;
+public class LifeVest {
+  private String surfaceColor;
 
   @Inject
-  Villa(Bed bed, MiniBar miniBar, TV tv, Phone phone, Pool pool, Tourist tourist) {
-    this.bed = bed;
-    this.miniBar = miniBar;
-    this.tv = tv;
-    this.phone = phone;
-    this.pool = pool;
-    this.tourist = tourist;
+  LifeVest() {
+    switch (Math.abs((new Random().nextInt()) % 5)) {
+      case 0:
+        surfaceColor = "red";
+        break;
+      case 1:
+        surfaceColor = "yellow";
+        break;
+      case 2:
+        surfaceColor = "blue";
+        break;
+      case 3:
+        surfaceColor = "green";
+        break;
+      case 4:
+        surfaceColor = "pink";
+        break;
+    }
   }
 }
