+++ project/src/com/google/daggerquery/example/Manager.java
@@ -14,12 +14,20 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-package com.google.daggerquery.example.models;
+package com.google.daggerquery.example;
 
 import javax.inject.Inject;
 
-public class TV {
+public class Manager implements Staff {
+  private final Phone phone;
+
   @Inject
-  TV() {
+  Manager(Phone phone) {
+    this.phone = phone;
+  }
+
+  @Override
+  public void help() {
+    System.out.println("Hotel Manager helped the tourist to organize an excursion.");
   }
 }
