+++ project/src/com/google/daggerquery/example/MiniBar.java
@@ -14,8 +14,19 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-package com.google.daggerquery.example.models;
+package com.google.daggerquery.example;
 
-public interface Staff {
-  void help();
+import javax.inject.Inject;
+
+public class MiniBar {
+  private final CocaCola cocaCola;
+  private final Fanta fanta;
+  private final Sprite sprite;
+
+  @Inject
+  MiniBar(CocaCola cocaCola, Fanta fanta, Sprite sprite) {
+    this.cocaCola = cocaCola;
+    this.fanta = fanta;
+    this.sprite = sprite;
+  }
 }
