+++ project/src/com/google/daggerquery/example/Pool.java
@@ -14,20 +14,19 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-package com.google.daggerquery.example.models;
+package com.google.daggerquery.example;
 
 import javax.inject.Inject;
 
-public class Lifeguard implements Staff {
-  private final LifeVest lifeVest;
+public class Pool {
+  private final InflatableDonut inflatableDonut;
+  private final InflatableFlamingo inflatableFlamingo;
+  private final InflatableUnicorn inflatableUnicorn;
 
   @Inject
-  Lifeguard(LifeVest lifeVest) {
-    this.lifeVest = lifeVest;
-  }
-
-  @Override
-  public void help() {
-    System.out.println("Lifeguard has helped a tourist to escape big waves 🌊.");
+  Pool(InflatableUnicorn unicorn, InflatableFlamingo flamingo, InflatableDonut donut) {
+    this.inflatableDonut = donut;
+    this.inflatableFlamingo = flamingo;
+    this.inflatableUnicorn = unicorn;
   }
 }
