+++ project/src/com/google/daggerquery/example/Staff.java
@@ -14,12 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-package com.google.daggerquery.example.models;
+package com.google.daggerquery.example;
 
-import javax.inject.Inject;
-
-public class Fanta {
-  @Inject
-  Fanta() {
-  }
+public interface Staff {
+  void help();
 }
