+++ project/src/com/google/daggerquery/example/TV.java
@@ -14,17 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-package com.google.daggerquery.example.models;
+package com.google.daggerquery.example;
 
 import javax.inject.Inject;
 
-public class Tourist {
-  private final Towel towel;
-  private final Phone phone;
-
+public class TV {
   @Inject
-  Tourist(Towel towel, Phone phone) {
-    this.towel = towel;
-    this.phone = phone;
+  TV() {
   }
 }
