+++ project/src/com/google/daggerquery/example/Tourist.java
@@ -14,23 +14,17 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-package com.google.daggerquery.example.models;
+package com.google.daggerquery.example;
 
 import javax.inject.Inject;
-import java.util.Random;
 
-public class Phone {
-  private String model;
+public class Tourist {
+  private final Towel towel;
+  private final Phone phone;
 
   @Inject
-  Phone() {
-    switch (Math.abs((new Random().nextInt()) % 2)) {
-      case 0:
-        model = "Google Pixel 4 XL";
-        break;
-      case 1:
-        model = "iPhone XS";
-        break;
-    }
+  Tourist(Towel towel, Phone phone) {
+    this.towel = towel;
+    this.phone = phone;
   }
 }
