--- project/src/com/google/daggerquery/example/modules/BUILD
@@ -1,28 +1,0 @@
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-load("@rules_java//java:defs.bzl", "java_library")
-load("@dagger//:workspace_defs.bzl", "dagger_rules")
-
-dagger_rules()
-package(default_visibility = ["//visibility:public"])
-
-java_library(
-    name = "example_modules",
-    srcs = glob(["*.java"]),
-    deps = [
-        "//src/com/google/daggerquery/example/models:example_models",
-        ":dagger",
-    ],
-)
